# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Music APIs",
    "version": "1.0.0",
    "description": "API documentation for musicApis project"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/playlists": {
      "post": {
        "summary": "Create a new playlist",
        "tags": ["Playlists"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "desc": {
                    "type": "string"
                  },
                  "img": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Playlist created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "summary": "Get all playlists",
        "tags": ["Playlists"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of playlists"
          }
        }
      }
    },
    "/playlists/favourite": {
      "get": {
        "summary": "Get playlists created by current user",
        "tags": ["Playlists"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's playlists"
          }
        }
      }
    },
    "/playlists/{id}": {
      "get": {
        "summary": "Get playlist by ID",
        "tags": ["Playlists"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist details"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a playlist",
        "tags": ["Playlists"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist deleted"
          },
          "403": {
            "description": "Not allowed"
          }
        }
      }
    },
    "/playlists/edit/{id}": {
      "put": {
        "summary": "Edit a playlist by ID",
        "tags": ["Playlists"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "desc": {
                    "type": "string"
                  },
                  "img": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Playlist updated"
          },
          "403": {
            "description": "Not allowed"
          }
        }
      }
    },
    "/playlists/add-song": {
      "put": {
        "summary": "Add a song to a playlist",
        "tags": ["Playlists"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["playlistId", "songId"],
                "properties": {
                  "playlistId": {
                    "type": "string"
                  },
                  "songId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Song added to playlist"
          }
        }
      }
    },
    "/playlists/remove-song": {
      "put": {
        "summary": "Remove a song from a playlist",
        "tags": ["Playlists"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["playlistId", "songId"],
                "properties": {
                  "playlistId": {
                    "type": "string"
                  },
                  "songId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Song removed from playlist"
          }
        }
      }
    },
...
  },
...
}